ice_breaker_app/requirements.txt
python-dotenv = "*"
langchain-core = "*"
langchain = "*"
langchain-openai = "*"
langchain-community = "*"
langchainhub = "*"
